Remove redundant cast

+++ server/src/test/java/org/realityforge/replicant/server/ChangeTest.java
@@ -62,7 +62,7 @@ public class ChangeTest
     change1.merge( change2 );
 
     assertEquals( change1.getChannels().size(), 2 );
-    assertEquals( change1.getChannels().get( 2 ), (Integer) 3 );
+    assertEquals( change1.getChannels().get( 2 ), 3 );
     assertEquals( change1.getEntityMessage().getAttributeValues().get( MessageTestUtil.ATTR_KEY1 ), "aZ" );
     assertEquals( change1.getEntityMessage().getRoutingKeys().get( MessageTestUtil.ROUTING_KEY2 ), "r3" );
   }
